fix(agent): keep Pi out of the dashboard bundle
agent-chat imported foldEvents from the agent-runtime package root, which re-exports the Pi
execution stack; pi-ai's env-api-keys module issues variable-specifier dynamic imports of node
builtins at load, which Turbopack cannot resolve and which surface as an unhandled rejection in
the dash SSR bundle. `@chia/agent-runtime/events` now points at wire/index.ts (schema, fold, clip —
no Pi dependency) and the dashboard imports from there.
